+++ include/smart_rtp/tcp/TcpServer.h
@@ -31,16 +31,16 @@ public:
 		std::thread *m_listener;
 		OnReceiveCallback *m_on_receive_callback;
 		void *m_cache;
-		size_t m_cache_size, m_connected_cache_size;
+		uint32_t m_cache_size, m_connected_cache_size;
 		bool m_listener_alive;
 		uint64_t m_last_received_time;
 	public:
 		Client(TcpServer *server_ptr, int32_t connection_sockfd, sockaddr *addr, OnReceiveCallback *);
 		virtual ~Client();
 		uint64_t id();
-		size_t cache_size();
-		size_t get_connection_cache_size();
-		void set_connection_cache_size(size_t);
+		uint32_t cache_size();
+		uint32_t get_connection_cache_size();
+		void set_connection_cache_size(uint32_t);
 		uint64_t last_receive_time();
 	private:
 		void listener_fn();
@@ -54,7 +54,7 @@ public:
 		virtual ~OnReceiveCallback() {};
 		virtual void onConnected(uint64_t client_id) = 0;
 		virtual void onDisconnected(uint64_t client_id) = 0;
-		virtual void onRequestToRepeat(Client *client, void *data, size_t size) = 0;
+		virtual void onRequestToRepeat(Client *client, void *data, uint32_t size) = 0;
 	};
 private:
 	int32_t m_self_sockfd;
@@ -72,16 +72,16 @@ public:
     TcpServer(int port);
     virtual ~TcpServer();
     ConnectionStatus start(OnReceiveCallback *callback);
-    void send(Client *client, void *data, size_t size);
+    void send(Client *client, void *data, uint32_t size);
     void close();
 
     static void print_status_name(const ConnectionStatus &status);
 
 
 private:
-    void onClientReceiveCallbackWrapper(Client *client, void *data, size_t size);
+    void onClientReceiveCallbackWrapper(Client *client, void *data, uint32_t size);
     void ping_checker_thread();
-    size_t get_used_cache_size();
+    uint32_t get_used_cache_size();
 };
